Pi5 serial support (#19)
Update the bootloader configuration to enable PiHat's that use pins 8/10 to be accessible via "/dev/serial0" on [pi3], [pi4], [pi5], and [pi0] devices.

Adds an "asl3-firstboot-complete" script/service that reboots the system after the initial (firstboot) actions including configuration and package updates have completed..

Adds a message-of-the-day note while the firstboot actions are active.